requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing and ML
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
torch==2.1.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Visualization
matplotlib==3.8.2
plotly==5.17.0
streamlit==1.29.0

# Utilities
schedule==1.2.0
python-telegram-bot==20.7
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0
